refactor(api): unify storage API to use direct S3 presigned URLs
- Add /api/storages/download endpoint returning presigned S3 GET URL
- Update CLI pull commands to download directly from S3
- Update CLI system-storage to use direct upload (prepare/commit flow)
- Remove legacy upload endpoints:
  - DELETE /api/storages (POST handler)
  - DELETE /api/webhooks/agent/storages
  - DELETE /api/webhooks/agent/storages/incremental
- Remove legacy sandbox scripts:
  - DELETE vas_snapshot.py.ts
  - DELETE incremental.py.ts
- Remove checkpoint.py.ts fallback to legacy upload methods
- Clean up unused env vars (ARTIFACT_MANIFEST_URL)

All storage operations now use consistent presigned URL pattern:
- Upload: prepare → S3 PUT → commit
- Download: get presigned URL → S3 GET
