--- turbo/apps/web/app/api/storages/__tests__/route.test.ts
@@ -1,280 +1,0 @@
-/**
- * @vitest-environment node
- */
-import {
-  describe,
-  it,
-  expect,
-  vi,
-  beforeEach,
-  beforeAll,
-  afterAll,
-} from "vitest";
-import { eq } from "drizzle-orm";
-import { initServices } from "../../../../src/lib/init-services";
-import { storages, storageVersions } from "../../../../src/db/schema/storage";
-
-// Mock external dependencies
-vi.mock("../../../../src/lib/auth/get-user-id", () => ({
-  getUserId: vi.fn().mockResolvedValue("test-user-storages-route"),
-}));
-
-vi.mock("../../../../src/lib/s3/s3-client", () => ({
-  uploadStorageVersionArchive: vi.fn().mockResolvedValue({
-    s3Prefix: "test-prefix",
-    filesUploaded: 0,
-    totalBytes: 0,
-    manifest: {},
-  }),
-  downloadS3Object: vi.fn().mockResolvedValue(undefined),
-}));
-
-vi.mock("../../../../src/lib/blob/blob-service", () => ({
-  blobService: {
-    uploadBlobs: vi.fn().mockResolvedValue({
-      hashes: new Map(),
-      newBlobsCount: 0,
-      existingBlobsCount: 0,
-    }),
-  },
-}));
-
-// Set required environment variables
-process.env.S3_USER_STORAGES_NAME = "test-storages-bucket";
-
-// Test constants
-const TEST_USER_ID = "test-user-storages-route";
-const TEST_PREFIX = "test-storage-route-";
-
-describe("Storages API Route", () => {
-  beforeAll(async () => {
-    initServices();
-  });
-
-  beforeEach(async () => {
-    vi.clearAllMocks();
-
-    // Clean up test data
-    await globalThis.services.db
-      .update(storages)
-      .set({ headVersionId: null })
-      .where(eq(storages.userId, TEST_USER_ID));
-
-    const testStorages = await globalThis.services.db
-      .select({ id: storages.id })
-      .from(storages)
-      .where(eq(storages.userId, TEST_USER_ID));
-
-    for (const storage of testStorages) {
-      await globalThis.services.db
-        .delete(storageVersions)
-        .where(eq(storageVersions.storageId, storage.id));
-    }
-
-    await globalThis.services.db
-      .delete(storages)
-      .where(eq(storages.userId, TEST_USER_ID));
-  });
-
-  afterAll(async () => {
-    // Final cleanup
-    await globalThis.services.db
-      .update(storages)
-      .set({ headVersionId: null })
-      .where(eq(storages.userId, TEST_USER_ID));
-
-    const testStorages = await globalThis.services.db
-      .select({ id: storages.id })
-      .from(storages)
-      .where(eq(storages.userId, TEST_USER_ID));
-
-    for (const storage of testStorages) {
-      await globalThis.services.db
-        .delete(storageVersions)
-        .where(eq(storageVersions.storageId, storage.id));
-    }
-
-    await globalThis.services.db
-      .delete(storages)
-      .where(eq(storages.userId, TEST_USER_ID));
-  });
-
-  describe("Empty Artifact Handling", () => {
-    describe("GET - 204 for empty artifacts", () => {
-      it("should return 204 No Content when fileCount is 0", async () => {
-        // Create storage with fileCount=0
-        const storageName = `${TEST_PREFIX}empty-artifact`;
-        const [storage] = await globalThis.services.db
-          .insert(storages)
-          .values({
-            userId: TEST_USER_ID,
-            name: storageName,
-            type: "artifact",
-            s3Prefix: `${TEST_USER_ID}/artifact/${storageName}`,
-            size: 0,
-            fileCount: 0,
-          })
-          .returning();
-
-        const versionId = `${TEST_PREFIX}empty-version`;
-        await globalThis.services.db.insert(storageVersions).values({
-          id: versionId,
-          storageId: storage!.id,
-          s3Key: `${TEST_USER_ID}/artifact/${storageName}/${versionId}`,
-          size: 0,
-          fileCount: 0,
-          createdBy: TEST_USER_ID,
-        });
-
-        // Update storage with head version
-        await globalThis.services.db
-          .update(storages)
-          .set({ headVersionId: versionId })
-          .where(eq(storages.id, storage!.id));
-
-        // Import GET handler after mocks are set up
-        const { GET } = await import("../route");
-
-        // Create request
-        const url = `http://localhost:3000/api/storages?name=${storageName}&type=artifact`;
-        const request = new Request(url, { method: "GET" });
-
-        // Call handler
-        const response = await GET(
-          request as unknown as import("next/server").NextRequest,
-        );
-
-        // Verify 204 response
-        expect(response.status).toBe(204);
-        expect(response.body).toBeNull();
-      });
-
-      it("should return 200 with tar.gz when fileCount > 0", async () => {
-        // Create storage with fileCount > 0
-        const storageName = `${TEST_PREFIX}nonempty-artifact`;
-        const [storage] = await globalThis.services.db
-          .insert(storages)
-          .values({
-            userId: TEST_USER_ID,
-            name: storageName,
-            type: "artifact",
-            s3Prefix: `${TEST_USER_ID}/artifact/${storageName}`,
-            size: 100,
-            fileCount: 5,
-          })
-          .returning();
-
-        const versionId = `${TEST_PREFIX}nonempty-version`;
-        await globalThis.services.db.insert(storageVersions).values({
-          id: versionId,
-          storageId: storage!.id,
-          s3Key: `${TEST_USER_ID}/artifact/${storageName}/${versionId}`,
-          size: 100,
-          fileCount: 5,
-          createdBy: TEST_USER_ID,
-        });
-
-        // Update storage with head version
-        await globalThis.services.db
-          .update(storages)
-          .set({ headVersionId: versionId })
-          .where(eq(storages.id, storage!.id));
-
-        // Import GET handler after mocks are set up
-        const { GET } = await import("../route");
-
-        // Create request
-        const url = `http://localhost:3000/api/storages?name=${storageName}&type=artifact`;
-        const request = new Request(url, { method: "GET" });
-
-        // This will fail because S3 download is mocked to not actually create a file
-        // But we can verify the handler doesn't return 204 for non-empty artifacts
-        // The actual file download would happen in integration tests
-        try {
-          const response = await GET(
-            request as unknown as import("next/server").NextRequest,
-          );
-          // If we get here, verify it's not a 204
-          expect(response.status).not.toBe(204);
-        } catch {
-          // Expected - S3 download mock doesn't create real file
-          // The important thing is that we didn't get a 204 response
-        }
-      });
-    });
-
-    describe("POST - Skip S3 for empty artifacts", () => {
-      it("should skip S3 upload when fileCount is 0", async () => {
-        // Import mocked functions
-        const { uploadStorageVersionArchive } = await import(
-          "../../../../src/lib/s3/s3-client"
-        );
-        const { blobService } = await import(
-          "../../../../src/lib/blob/blob-service"
-        );
-
-        // Import POST handler
-        const { POST } = await import("../route");
-
-        // Create an empty tar.gz file for testing
-        const tar = await import("tar");
-        const fs = await import("node:fs");
-        const path = await import("node:path");
-        const os = await import("node:os");
-
-        const tempDir = path.join(
-          os.tmpdir(),
-          `test-empty-upload-${Date.now()}`,
-        );
-        await fs.promises.mkdir(tempDir, { recursive: true });
-        const emptyDir = path.join(tempDir, "empty");
-        await fs.promises.mkdir(emptyDir, { recursive: true });
-        const tarPath = path.join(tempDir, "empty.tar.gz");
-
-        await tar.create(
-          {
-            gzip: true,
-            file: tarPath,
-            cwd: emptyDir,
-          },
-          ["."],
-        );
-
-        const tarBuffer = await fs.promises.readFile(tarPath);
-
-        // Create FormData with empty tar.gz
-        const formData = new FormData();
-        formData.append("name", `${TEST_PREFIX}empty-upload`);
-        formData.append("type", "artifact");
-        formData.append(
-          "file",
-          new Blob([new Uint8Array(tarBuffer)], { type: "application/gzip" }),
-          "empty.tar.gz",
-        );
-
-        const request = new Request("http://localhost:3000/api/storages", {
-          method: "POST",
-          body: formData,
-        });
-
-        // Call handler
-        const response = await POST(
-          request as unknown as import("next/server").NextRequest,
-        );
-        const responseJson = await response.json();
-
-        // Verify response indicates empty artifact
-        expect(responseJson.fileCount).toBe(0);
-
-        // Verify S3 upload was NOT called (skipped for empty artifacts)
-        expect(uploadStorageVersionArchive).not.toHaveBeenCalled();
-
-        // Verify blob upload was NOT called (skipped for empty artifacts)
-        expect(blobService.uploadBlobs).not.toHaveBeenCalled();
-
-        // Cleanup
-        await fs.promises.rm(tempDir, { recursive: true, force: true });
-      });
-    });
-  });
-});
